# Fix: lakectl local diff slowness