Closes #7643

## Change Description

### Background

lakectl local had a bug due to the different way standard FS lists files recursively in comparison with object stores (S3).
A fix was introduced which used the local adapter file walker to be used also for lakectl local, in order to diff between local and remote.
Due to the mentioned change, the local diff performance was greatly impacted.

### Bug Fix

Improve local diff performance by creating a new walk functions that simulates the order in which object stores list files.
Relying on the FS walk implementation and modifying the behavior when iterating over directories

### Testing Details

Manually testing the scenario of ~500K file diff. Operation time was reduced **from 18s to 3s**
Existing test verify correctness was not broken

### Breaking Change?

No

## Additional info

**before:**
```
lakectl local status -l  17.47s user 2.41s system 256% cpu 7.765 total
```

**after:**
```
lakectl local status -l  2.97s user 1.00s system 94% cpu 4.212 total
```