@N-o-Z 2024-06-09
> Not sure about this code. We should find a way to test this very precisely.

Thanks for the review.
I understand the concern, however I'm missing what exactly is missing in terms of confidence.
I've added additional unit tests for the walk function itself. The flows are extensively tested due to the recent work we did with bug fixes. I've also manually verified the performance improvement and documented it in the description.
If there are any other specific concerns - please let me know