@N-o-Z 2024-06-11
> I believe further testing is required. So I pushed some fuzzing code; I think it found a bug. I guess we need to fix either the fuzzer or the code. But I would like us to have _some_ fuzzing here.

@arielshaqed Thanks for the test! I'm reverting the previous change as this is the root cause of the bug!